requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0